requirements.txt
PYJWT
colorama